+++ qase-cucumberjs/src/models.ts
@@ -10,6 +10,7 @@ export interface TestMetadata {
   parameters: Record<string, string>;
   group_params: Record<string, string>;
   suite: string | null;
+  tags: string[];
 }
 
 export interface ScenarioData {